requirements.txt
python-binance
fastapi
pydantic
uvicorn
python-dotenv